Experiment with plugins extending core classes
The 'resources/basic' plugin augments Node with resource attributes,
namely CPUs, GPUs and memory.  This mostly serves as an exploration of
how a GRES plugin might work.
