+++ lib/flight_scheduler/plugins/resources.rb
@@ -1,6 +1,13 @@
+require 'active_model'
+require 'forwardable'
+
 module FlightScheduler
   class Plugins
     class Resources
+      def self.init
+        ::Node.prepend(NodeExtensions)
+      end
+
       class EventProcessor
         def daemon_connected(node, message)
           # Update the node's attributes
@@ -28,11 +35,55 @@ module FlightScheduler
         end
       end
 
+      class NodeAttributes
+        include ActiveModel::Model
+
+        DELEGATES = [:cpus, :gpus, :memory]
+
+        attr_writer :type
+        attr_accessor(*DELEGATES)
+        validates(*DELEGATES, allow_nil: true, numericality: { only_integers: true })
+
+        def type
+          str = @type.to_s
+          str.empty? ? 'unknown' : str
+        end
+
+        def to_h
+          self.class::DELEGATES.each_with_object({ type: type }) do |key, memo|
+            memo[key] = self.send(key)
+          end
+        end
+
+        def ==(other)
+          return false unless other.class == self.class
+          self.class::DELEGATES.all? do |key|
+            self.send(key) == other.send(key)
+          end
+        end
+      end
+
+      module NodeExtensions
+        extend Forwardable
+        attr_accessor :attributes
+        def_delegators  :attributes, :type, *NodeAttributes::DELEGATES
+
+        def initialize(*args)
+          super
+          @attributes = NodeAttributes.new(cpus: 1, memory: 1048576)
+        end
+      end
+
       def event_processor
         @event_processor ||= EventProcessor.new
       end
+
+      def resources_for(node)
+        node.attributes.to_h
+      end
     end
 
+    Resources.init
     FlightScheduler.app.plugins.register('resources/basic', Resources.new)
   end
 end
